requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-recaptcha3==0.4.0
django_debug_toolbar==3.8.1
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.5
Pillow==9.4.0
psycopg2-binary==2.9.5
python-decouple==3.7
pytils==0.4.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
